Refactor: replace Double with BigDecimal, fix DI, exception handling, and REST mappings

+++ src/main/java/com/vendingMachine/model/dto/PurchaseRequest.java
@@ -3,13 +3,13 @@ package com.vendingMachine.model.dto;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
 @Builder
 public class PurchaseRequest {
 
-    private List<Double> listCoins;
+    private List<BigDecimal> listCoins;
 
-    
 }
